archive/automated_notebook/pyproject.toml
[tool.poetry]
name = "automated_notebook"
version = "0.1.0"
description = ""
authors = ["David Aronchick <aronchick@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8"
jupyter = "^1.0.0"
ipython = "^7.22.0"
dill = "^0.3.3"
beautifulsoup4 = "^4.9.3"
Pillow = "^8.2.0"
plotly = "^4.14.3"
tensorflow = "^2.4.1"
pandas = "^1.2.4"
chart-studio = "^1.1.0"
matplotlib = "^3.4.1"
scipy = {version = "^1.6.3", python = ">=3.8,<3.10"}
jupytext = {version = "^1.11.2", python = "3.8"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
